Finish q_sort's merge() and modify re_tail,head
q_remove_tail() cannot pass the time complexity test,
because its time complexity is in O(n) not in constant time.
So I modify it and q_remove_head with removing the use of strlen().

q_sort's merge() was finished with the use of pointer to pointer.
The detail description can be seen in
https://hackmd.io/@_QH1V5HYRWitWQueAZDAjQ/HkdhzXYJ5